requirements.txt
django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
slack_sdk==3.33.1
sib-api-v3-sdk==7.6.0
Pillow==10.4.0

